Fix compiler warnings: Unused functions
Move function definitions in include/blisp_util.h to lib/blisp_util.c.

Defining them as static leads to internal linkage, accessible in that
translation unit, hence the warning. Their references are to be handled
in other translation units, so they shouldn't be static.
