+++ include/blisp_util.h
@@ -2,35 +2,18 @@
 #ifndef _BLISP_UTIL_H
 #define _BLISP_UTIL_H
 
-#include <stdarg.h>
-#include <stdio.h>
+#include <stdint.h>
 #ifdef WIN32
-#include <windows.h>
+#  include <windows.h>
 #else
-#include <time.h>
+#  include <time.h>
 #endif
 
-static void blisp_dlog(const char* format, ...)
-{
-  fflush(stdout);
-  va_list args;
-  va_start(args, format);
-  vfprintf(stderr, format, args);
-  va_end(args);
-  fputc('\n', stderr);
-}
+void blisp_dlog(const char* format, ...);
 
+void sleep_ms(int milliseconds);
 
-static void sleep_ms(int milliseconds) {
-#ifdef WIN32
-  Sleep(milliseconds);
-#else
-  struct timespec ts;
-  ts.tv_sec = milliseconds / 1000;
-  ts.tv_nsec = (milliseconds % 1000) * 1000000;
-  nanosleep(&ts, NULL);
-#endif
-}
+uint32_t crc32_calculate(const void *data, size_t data_len);
 
 /**
  * * Generated on Mon Jan  9 19:56:36 2023
@@ -79,20 +62,5 @@ static const uint32_t crc_table[256] = {
     0xbdbdf21c, 0xcabac28a, 0x53b39330, 0x24b4a3a6, 0xbad03605, 0xcdd70693, 0x54de5729, 0x23d967bf,
     0xb3667a2e, 0xc4614ab8, 0x5d681b02, 0x2a6f2b94, 0xb40bbe37, 0xc30c8ea1, 0x5a05df1b, 0x2d02ef8d
 };
-
-static uint32_t crc32_calculate(const void *data, size_t data_len)
-{
-  uint32_t crc = 0xffffffff;
-  const unsigned char *d = (const unsigned char *)data;
-  unsigned int tbl_idx;
-
-  while (data_len--) {
-    tbl_idx = (crc ^ *d) & 0xff;
-    crc = (crc_table[tbl_idx] ^ (crc >> 8)) & 0xffffffff;
-    d++;
-  }
-  return (crc & 0xffffffff) ^ 0xffffffff;
-}
-
 
 #endif
